Hybrid and MyHybrid should now be identical

+++ Algorithms/Base/BaseSimilarityMatrixRecommender.py
@@ -84,7 +84,7 @@ class BaseItemSimilarityMatrixRecommender(BaseSimilarityMatrixRecommender):
         user_profile_array = self.URM_train[user_id_array]
 
         if items_to_compute is not None:
-            item_scores = - np.ones((len(user_id_array), self.URM_train.shape[1]), dtype=np.float32)*np.inf
+            item_scores = - np.ones((len(user_id_array), self.URM_train.shape[1]), dtype=np.float64)*np.inf
             item_scores_all = user_profile_array.dot(self.W_sparse).toarray()
             item_scores[:, items_to_compute] = item_scores_all[:, items_to_compute]
         else:
@@ -108,7 +108,7 @@ class BaseUserSimilarityMatrixRecommender(BaseSimilarityMatrixRecommender):
         user_weights_array = self.W_sparse[user_id_array]
 
         if items_to_compute is not None:
-            item_scores = - np.ones((len(user_id_array), self.URM_train.shape[1]), dtype=np.float32)*np.inf
+            item_scores = - np.ones((len(user_id_array), self.URM_train.shape[1]), dtype=np.float64)*np.inf
             item_scores_all = user_weights_array.dot(self.URM_train).toarray()
             item_scores[:, items_to_compute] = item_scores_all[:, items_to_compute]
         else:
